#50 refactor: actor-settings/actor/game-clock の index.ts から types/constants の re-export を削除
- aggregation-index.md ルール改訂に合わせ、各 index.ts は実装 (store.ts) と hook/context
  (context.tsx) のみを re-export する形にする
- component・_lib からの型・定数参照は種類別ファイル (constants.ts/types.ts) へ直接に変更
