+++ src/prototypes/time-control/time-control-03/__tests__/actor-settings-store.test.ts
@@ -1,7 +1,5 @@
-import {
-  createActorSettingsStore,
-  DEFAULT_ACTOR_SETTINGS,
-} from '../_stores/actor-settings'
+import { createActorSettingsStore } from '../_stores/actor-settings'
+import { DEFAULT_ACTOR_SETTINGS } from '../_stores/actor-settings/constants'
 
 test('未設定 actor の設定はデフォルト値', () => {
   const store = createActorSettingsStore()
