+++ src/prototypes/time-control/time-control-03/_components/current-position-marker/index.tsx
@@ -4,7 +4,8 @@ import { getTickMs } from '../../../time-control-02/_lib/get-tick-ms'
 import { usePositionStoreApi } from '../../_contexts/position-store-context'
 import { useStageTransform } from '../../_contexts/stage-transform-context'
 import { toPixelStyle } from '../../_lib/stage-coords'
-import { DEFAULT_ACTOR_INFO, useActorStore } from '../../_stores/actor'
+import { useActorStore } from '../../_stores/actor'
+import { DEFAULT_ACTOR_INFO } from '../../_stores/actor/constants'
 import { DEFAULT_POSITION } from '../../_stores/position-store'
 import { ActorId, Position } from '../../types'
