+++ src/prototypes/time-control/time-control-03/_stores/actor-settings/index.ts
@@ -1,4 +1,2 @@
-export * from './constants'
 export * from './context'
 export * from './store'
-export * from './types'
