+++ src/prototypes/time-control/time-control-03/_stores/game-clock/index.ts
@@ -1,3 +1,2 @@
 export * from './context'
 export * from './store'
-export * from './types'
